refactor:소켓 서비스 로직 분리(backend)
기존의 웹소켓 로직을 loader와 service로 분리합니다
