+++ back/package.json
@@ -24,6 +24,7 @@
         "@types/jest": "^27.0.3",
         "@types/node": "^16.11.7",
         "@types/node-fetch": "^2.5.12",
+        "@types/ws": "^8.2.1",
         "@typescript-eslint/eslint-plugin": "^4.33.0",
         "@typescript-eslint/parser": "^4.33.0",
         "better-sqlite3": "^7.4.5",
